requirements.txt
asgiref==3.8.1
distlib==0.3.8
Django==5.0.7
djangorestframework==3.15.2
filelock==3.15.4
platformdirs==4.2.2
sqlparse==0.5.1
virtualenv==20.26.3
